requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22.2,<2.0.0

# Image processing
Pillow>=9.0.0
opencv-python-headless>=4.5.0

# Data processing
pandas>=1.5.0

# AWS Lambda support (optional)
mangum>=0.17.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Environment and configuration
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.0.0

# Type checking
mypy>=1.0.0
types-Pillow>=9.0.0